+++ Test/UnitTests/TestDataLayer/TestSqlServerHelpers.cs
@@ -1,7 +1,12 @@
 ﻿// Copyright (c) 2017 Jon P Smith, GitHub: JonPSmith, web: http://www.thereformedprogrammer.net/
 // Licensed under MIT licence. See License.txt in the project root for license information.
 
+using System.Collections.Generic;
+#if NETCOREAPP2_1
 using System.Data.SqlClient;
+#elif NETCOREAPP3_0
+using Microsoft.Data.SqlClient;
+#endif
 using System.Linq;
 using DataLayer.EfCode.BookApp;
 using Microsoft.EntityFrameworkCore;
@@ -98,16 +103,16 @@ namespace Test.UnitTests.TestDataLayer
         public void TestCreateDbToGetLogsOk()
         {
             //SETUP
-            var options = this.CreateUniqueMethodOptions<BookContext>();
+            var logs = new List<LogOutput>();
+            var options = this.CreateUniqueClassOptionsWithLogging<BookContext>(log => logs.Add(log));
             using (var context = new BookContext(options))
             {
                 //ATTEMPT
-                var logs = context.SetupLogging();
                 context.Database.EnsureDeleted();
                 context.Database.EnsureCreated();
 
                 //VERIFY
-                foreach (var log in logs.ToList())
+                foreach (var log in logs)
                 {
                     _output.WriteLine(log.ToString());
                 }