+++ TestSupport/EfHelpers/WipeDbViaSql.cs
@@ -34,9 +34,11 @@ namespace TestSupport.EfHelpers
             foreach (var tableName in
                 context.GetTableNamesInOrderForWipe(addBracketsAndSchema, maxDepth, excludeTypes))
             {
-                context.Database
-                    .ExecuteSqlCommand(
-                        "DELETE FROM " + tableName);
+#if NETSTANDARD2_0
+                context.Database.ExecuteSqlCommand("DELETE FROM " + tableName);
+#elif NETSTANDARD2_1
+                context.Database.ExecuteSqlRaw("DELETE FROM " + tableName);
+#endif
             }
         }
 
@@ -49,7 +51,8 @@ namespace TestSupport.EfHelpers
         {
             var allEntities = context.Model
                 .GetEntityTypes()
-                .Where(x => !excludeTypes.Contains(x.ClrType))
+                .Where(x => !excludeTypes.Contains(x.ClrType) 
+                            && x.FindPrimaryKey() != null) //This excludes red-only entities
                 .ToList(); //#C
 
             ThrowExceptionIfCannotWipeSelfRef(allEntities); //#D
@@ -126,6 +129,7 @@ namespace TestSupport.EfHelpers
 
         private static string FormTableNameWithSchema(IEntityType entityType, bool addBracketsAndSchema)
         {
+#if NETSTANDARD2_0
             var relational = entityType.Relational();
             return addBracketsAndSchema 
                 ? "[" + (relational.Schema == null
@@ -133,6 +137,14 @@ namespace TestSupport.EfHelpers
                        : relational.Schema + "].[")
                    + relational.TableName + "]"
                 : relational.TableName;
+#elif NETSTANDARD2_1
+                        return addBracketsAndSchema 
+                ? "[" + (entityType.GetSchema() == null
+                       ? ""
+                       : entityType.GetSchema() + "].[")
+                   + entityType.GetTableName() + "]"
+                : entityType.GetTableName();
+#endif
         }
 
         private static void ThrowExceptionIfCannotWipeSelfRef(List<IEntityType> allEntities)